requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
requests
python-telegram-bot==13.15
